pyproject.toml
[project]
name = "operate-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hamza Farhan", email = "thehamza96@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "faker>=37.1.0",
    "fastapi[standard]>=0.115.12",
    "fastparquet>=2024.11.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.6",
    "loguru>=0.7.3",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "pydantic-ai>=0.0.55",
    "xlsxwriter>=3.2.2",
    "python-dateutil>=2.9.0.post0",
    "scipy>=1.15.2",
]

[project.scripts]
operate-ai = "operate_ai:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
